requirements.txt
fastapi
instructor[cerebras_cloud_sdk]
openai
pydantic
uvicorn
pytest
modal
arize-phoenix
arize-phoenix-otel
black
flake8
openinference-instrumentation-groq
openinference-instrumentation-openai
openinference-instrumentation-openai-agents
groq
sentry-sdk
braintrust
sqlalchemy
google-genai
openai-agents[litellm]